DevteroFlex: Verilator Full System Simulator infrastructure for regression tests
This is a private project that was merged to the main branch.
Here we developped the toolchain for full system simulation using the
API written in C.
With this toolchain we have a regression tests for multiple layers: MMU transactions, Transplants, Page Faults, Memory , executing a single instruction, and more.