Pipeline grants permission to MMU for flushing
To ensure that no dirty address can be accessed followed by a page
eviction, the pipeline must complete all pending cache accesses before
evicting the blocks.
We add a flush controller which ensures that once a flush request is
received from the MMU, the pipeline stops translating new addresses
and waits for all cache accesses to complete before allowing the MMU
to perform the flush.