Backpressure support removing queues size requirement of NB_THREADS
In our base design to remove the backpressure logic requirement we sized
all our queuest according to the maximum instructions in flight, which
is max NB_THREADs. This is fine in case we support a small number of
threads but given that we aim to support 128 threads we need a solution.
The solution is this credit system that keeps tracks how many entries
are left for the transaction receiver by peeking into transactions into
the memory and out of the queue.